HotelWise/CloudFunctions/ML_HTTP/requirements.txt
functions-framework==3.*
google-cloud-storage
pandas
fsspec
gcsfs
scikit-learn
pyarrow
flask